% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-SE-object.R
\name{create_summarized_experiment_object}
\alias{create_summarized_experiment_object}
\title{Creates a Summarized Experiment Object}
\usage{
create_summarized_experiment_object(GEOID, summary)
}
\arguments{
\item{geo_selection_df}{A dataframe containing the MPRA data for a user defined GEO identifier available in the legacy database.}
}
\description{
This function creates a Summarized Experiment Object containing Massively Parallel Reporter Assay (MPRA) data.
}
\examples{
summary_df <- read.table("../inst/summary.csv", header=T, sep=',') # Load summary file
geo_selection_df <- filter(summary_df, GEO_number=="GSE83894") # User select GEO identifier
se <- ExportSummarizedExperiment(geo_selection_df) # Create the Summarized Experimen
}
