% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-SE-object.R
\name{MPRAsnakeflow.ExportObject.SummarizedExperiment}
\alias{MPRAsnakeflow.ExportObject.SummarizedExperiment}
\title{This function creates a Summarized Experiment Object containing Massively Parallel Reporter Assay (MPRA) data.}
\usage{
MPRAsnakeflow.ExportObject.SummarizedExperiment(
  genome_build,
  genomic_note,
  metadata,
  reference_fasta,
  assigned_counts_replicate_list
)
}
\arguments{
\item{reference_fasta}{string path}
}
\description{
This function creates a Summarized Experiment Object containing Massively Parallel Reporter Assay (MPRA) data.
}
